requirements.txt
# requirements.txt
# 氢电混动车智能监督系统依赖

# 核心依赖
numpy>=1.20.0

# 串口通信
pyserial>=3.5

# 配置文件解析
pyyaml>=6.0

# 可选：如果需要 UDP 通信的高级功能
# 标准库已包含 socket，无需额外安装